requirements.txt
python-dotenv
gitpython
requests
